drop pip-shims as a dependency
