+++ news/132.bugfix
@@ -1,0 +1,1 @@
+Drop `pip-shims` package as a dependency.
