+++ pdm/models/caches.py
@@ -3,11 +3,11 @@ import json
 from pathlib import Path
 from typing import TYPE_CHECKING, Any, Dict, Optional
 
-import pip_shims
 from pip._vendor import requests
 
 from pdm._types import CandidateInfo
 from pdm.exceptions import CorruptedCacheError
+from pdm.models import pip_shims
 from pdm.utils import open_file
 
 if TYPE_CHECKING:
@@ -92,7 +92,7 @@ class HashCache(pip_shims.SafeFileCache):
         # to store it.
         hash_value = self.get(link.url)
         if not hash_value:
-            if link.hash:
+            if link.hash and link.hash_name in pip_shims.STRONG_HASHES:
                 hash_value = f"{link.hash_name}:{link.hash}"
             else:
                 hash_value = self._get_file_hash(link)
