+++ pyproject.toml
@@ -28,18 +28,17 @@ Documentation = "https://pdm.fming.dev"
 appdirs = "*"
 click = "*"
 distlib = "*"
-pip = ">=20.1"
-pip_shims = "*"
-pythonfinder = "*"
-tomlkit = "*"
-python-cfonts = "*"
-resolvelib = "<1.0.0,>=0.3.0"
-pdm-pep517 = "<1.0.0,>=0.1.0"
 importlib-metadata = {version = "*", marker = "python_version<'3.8'"}
+pdm-pep517 = "<1.0.0,>=0.1.0"
 pep517 = "*"
+pip = ">=20.1,<21"
 pycomplete = "<1.0.0,>=0.2.0"
+python-cfonts = "*"
 python-dotenv = "<1.0.0,>=0.15.0"
+pythonfinder = "*"
+resolvelib = "<1.0.0,>=0.3.0"
 shellingham = "<2.0.0,>=1.3.2"
+tomlkit = "*"
 
 [tool.pdm.dev-dependencies]
 pytest = "*"
